feat(generating dynamic tests): generating dynamic tests
