+++ src/generators/index.js
@@ -8,24 +8,24 @@ module.exports = async (tools) => {
 
   return {
     new: {
-      packageJson: await require('./new/packagejson')(tools),
-      entities: await require('./new/domain/entities')(tools),
-      connection: await require('./new/infra/database/connection')(tools),
-      repositories: await require('./new/infra/database/repositories/repositories')(tools),
-      migrations: options.postgres ? await require('./new/infra/database/migrations/migrations')(tools) : ignore,
-      useCases: await require('./new/domain/useCases')(tools),
-      useCasesTests: (options.entities === '') ? await require('./new/domain/useCasesTests')(tools) : ignore,
-      graphql: options.graphql ? await require('./new/infra/graphql')(tools) : ignore,
-      rest: options.rest ? await require('./new/infra/rest')(tools) : ignore,
-      config: infra ? await require('./new/infra/config')(tools) : ignore,
-      server: infra ? await require('./new/infra/server')(tools) : ignore,
-      index: await require('./new/index')(tools)
+      packageJson: await require('./src/packagejson')(tools),
+      entities: await require('./src/domain/entities')(tools),
+      connection: await require('./src/infra/database/connection')(tools),
+      repositories: await require('./src/infra/database/repositories/repositories')(tools),
+      migrations: options.postgres ? await require('./src/infra/database/migrations/migrations')(tools) : ignore,
+      useCases: await require('./src/domain/usecases/useCases')(tools),
+      useCasesTests: (options.entities === '') ? await require('./src/domain/usecases/unitTests')(tools) : ignore,
+      graphql: options.graphql ? await require('./src/infra/graphql')(tools) : ignore,
+      rest: options.rest ? await require('./src/infra/rest')(tools) : ignore,
+      config: infra ? await require('./src/infra/config')(tools) : ignore,
+      server: infra ? await require('./src/infra/server')(tools) : ignore,
+      index: await require('./src/index')(tools)
     },
     update: {
-      entities: await require('./new/domain/entities')(tools, true),
-      useCases: await require('./new/domain/useCases')(tools, true),
-      migrations: updateMigration ? await require('./new/infra/database/migrations/migrations')(tools) : ignore,
-      repositories: await require('./new/infra/database/repositories/repositories')(tools, true)
+      entities: await require('./src/domain/entities')(tools, true),
+      useCases: await require('./src/domain/usecases/useCases')(tools, true),
+      migrations: updateMigration ? await require('./src/infra/database/migrations/migrations')(tools) : ignore,
+      repositories: await require('./src/infra/database/repositories/repositories')(tools, true)
     }
   }
 }
