+++ src/generators/src/domain/entities.js
