+++ src/generators/src/infra/api/rest.js
