+++ src/generators/src/infra/config/index.js
