+++ src/generators/src/infra/rest.js
