+++ src/generators/src/packagejson.js
