+++ src/templates/domain/useCases/tests/getById.test.ejs
@@ -3,9 +3,7 @@ const assert = require('assert')
 const { <%= props.name.pascalCase %> } = require('../../entities')
 
 describe('Find one <%= props.name.pascalCase %>', () => {
-  function aUser({ hasAccess }) {
-    return { hasAccess }
-  }
+  const authorizedUser = { hasAccess = true }
 
   describe('Valid <%= props.name.pascalCase %>', () => {
 
@@ -18,18 +16,16 @@ describe('Find one <%= props.name.pascalCase %>', () => {
           async  findByID(id) { return ([retInjection]) }
         })
       }
-      const user = aUser({ hasAccess: true })
-      /*{ <%= props.request %> }*/
-      const req = { id: 5 }
+
+      const req = <%= props.request.valid %>
 
       // When
       const uc = getById<%= props.name.pascalCase %>(injection)()
-      await uc.authorize(<%= props.name.camelCase %>)
+      await uc.authorize(authorizedUser)
       const ret = await uc.run(req)
 
       // Then
       assert.ok(ret.isOk)      
-      assert.strictEqual(ret.ok.nickname, 'herbsUser')
       assert.strictEqual(ret.ok.isValid(), true)
 
     })
@@ -40,18 +36,16 @@ describe('Find one <%= props.name.pascalCase %>', () => {
     it('should not return if input is invalid', async () => {
       // Given
       const injection = {}
-      const user = aUser({ hasAccess: true })
-      /*{ <%= props.request %> }*/
-      const req = { id : '5' }
+
+      const req = <%= props.request.invalid %>
 
       // When
       const uc = getById<%= props.name.pascalCase %>(injection)()
-      await uc.authorize(<%= props.name.camelCase %>)
+      await uc.authorize(authorizedUser)
       const ret = await uc.run(req)
 
       // Then
       assert.ok(ret.isErr)
-      assert.deepStrictEqual(ret.err, {request :[{id:[{wrongType:"Number"}]}]})
     })
   })
 })
