+++ src/templates/domain/useCases/tests/update.test.ejs
@@ -4,9 +4,7 @@ const { <%= props.name.pascalCase %> } = require('../../entities')
 
 
 describe('Update <%= props.name.pascalCase %>', () => {
-  function aUser({ hasAccess }) {
-    return { hasAccess }
-  }
+  const authorizedUser = { hasAccess = true }
 
   describe('Valid <%= props.name.pascalCase %>', () => {
 
@@ -19,13 +17,12 @@ describe('Update <%= props.name.pascalCase %>', () => {
           async update(id) { return true }
         })
       }
-      const user = aUser({ hasAccess: true })
-      /*{ <%= props.request %> }*/
-      const req = {  id: 5, nickname: 'herbsUser', password: 'V&eryStr0ngP@$$'}
+
+      const req = <%= props.request.valid %>
 
       // When
       const uc = update<%= props.name.pascalCase %>(injection)()
-      await uc.authorize(<%= props.name.camelCase %>)
+      await uc.authorize(authorizedUser)
       const ret = await uc.run(req)
 
       // Then
@@ -39,18 +36,15 @@ describe('Update <%= props.name.pascalCase %>', () => {
     it('should not update invalid <%= props.name.pascalCase %>', async () => {
       // Given
       const injection = {}
-      const user = aUser({ hasAccess: true })
-      /*{ <%= props.request %> }*/
-      const req = {  nickname: 'herbsUser', password: 96587422 }
+      const req = <%= props.request.invalid %>
 
       // When
       const uc = update<%= props.name.pascalCase %>(injection)()
-      await uc.authorize(<%= props.name.camelCase %>)
+      await uc.authorize(authorizedUser)
       const ret = await uc.run(req)
 
       // Then
       assert.ok(ret.isErr)
-      assert.deepStrictEqual(ret.err, {request :[{password:[{wrongType:"String"}]}]})
     })
   })
 })
